fix: move CollectionSort to /shared and use boxicon
