+++ src/common/index.ts
@@ -3,7 +3,6 @@ export { default as ButtonProps } from "./Button"
 export { default as CalloutProps } from "./Callout"
 export { default as CardsProps } from "./Cards"
 export { default as CollectionCardProps } from "./CollectionCard"
-export { default as CollectionSortProps } from "./CollectionSort"
 export { default as ContentProps } from "./Content"
 export { default as ContentPageHeaderProps } from "./ContentPageHeader"
 export { default as FooterProps } from "./Footer"
