+++ src/templates/next/layouts/Collection/Collection.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react"
 import { CollectionCardProps } from "~/common"
-import { SortDirection, SortKey } from "~/common/CollectionSort"
+import { SortDirection, SortKey } from "~/templates/next/types/CollectionSort"
 import type { CollectionPageSchema } from "~/engine"
 import { CollectionCard } from "../../components"
-import CollectionSort from "../../components/CollectionSort"
+import CollectionSort from "../../components/shared/CollectionSort"
 import { Heading } from "../../typography/Heading"
 import { Paragraph } from "../../typography/Paragraph"
 import { Skeleton } from "../Skeleton"
